all basec functions done

+++ domain/src/main/java/com/ptitsn/domain/model/Weather.kt
@@ -3,5 +3,5 @@ package com.ptitsn.domain.model
 import java.util.*
 
 data class Weather(val city: CharSequence,
-                   val celsius: Float,
+                   val celsius: Double,
                    val date: Date)
