+++ presentation/src/main/java/com/ptitsn/presentation/ui/mapper/UiModelMapper.kt
@@ -21,7 +21,7 @@ open class UiModelMapper @Inject constructor(val resoruces: Resources) {
         return sdf.format(date)
     }
 
-    fun formateTocelsi(ce: Float): CharSequence {
+    fun formateTocelsi(ce: Double): CharSequence {
         return "$ce"
     }
